rename the include directory to vdk_signals
This allows to use #include <vdk_signals/signals.h> rather
than the possibly conflicting #include <signals.h> because
it is too generic.
